refactor: `Category` 컴포넌트 -> `RankingCategory`, `RegisterCategory` 컴포넌트로 분리
