+++ src/drawer/hooks/useGetNewRelease.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import { useRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
 import { getNewRelease } from '@/drawer/apis/getNewRelease';
 import { CategoryState } from '@/drawer/recoil/CategoryState';
@@ -8,7 +8,7 @@ import { ProductListResponse } from '@/drawer/types/ProductList.type';
 
 export const useGetNewRelease = () => {
   const [newReleases, setNewReleases] = useState<ProductListResponse['productList']>([]);
-  const [selectedCategory, setSelectedCategory] = useRecoilState(CategoryState);
+  const selectedCategory = useRecoilValue(CategoryState);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -24,5 +24,5 @@ export const useGetNewRelease = () => {
     console.log(newReleases);
   }, [selectedCategory]);
 
-  return { newReleases, selectedCategory, setSelectedCategory };
+  return { newReleases };
 };
